requirements.txt
fastapi==0.115.0
pydantic==2.9.2
langchain-huggingface==0.1.0
requests==2.28.2
python-dotenv==1.0.1
uvicorn==0.31.1
